machine learning/rl_starter/computer_vision/requirements.txt
torch>=2.1
torchvision>=0.16
matplotlib>=3.8
tqdm>=4.66
scikit-learn>=1.4
umap-learn>=0.5
